+++ CHANGES.md
@@ -8,9 +8,11 @@ Improvements 🙌:
  - Split network request `/keys/query` into smaller requests (250 users max) (#2925)
  - Crypto improvement | Bulk send NO_OLM withheld code
  - Display the room shield in all room setting screens
+ - Improve message with Emoji only detection (#3017)
 
 Bugfix 🐛:
- -
+ - Fix bad theme change for the MainActivity
+ - Handle encrypted reactions (#2509)
 
 Translations 🗣:
  -