+++ CHANGES.md
@@ -9,6 +9,7 @@ Improvements 🙌:
  - Crypto improvement | Bulk send NO_OLM withheld code
  - Display the room shield in all room setting screens
  - Improve message with Emoji only detection (#3017)
+ - Api interceptor to allow app developers peek responses (#2986)
 
 Bugfix 🐛:
  - Fix bad theme change for the MainActivity