+++ CHANGES.md
@@ -10,10 +10,12 @@ Improvements 🙌:
  - Display the room shield in all room setting screens
  - Improve message with Emoji only detection (#3017)
  - Picture preview when replying. Also add the image preview in the message detail bottomsheet (#2916)
+ - Api interceptor to allow app developers peek responses (#2986)
 
 Bugfix 🐛:
  - Fix bad theme change for the MainActivity
  - Handle encrypted reactions (#2509)
+ - Disable URL preview for some domains (#2995)
 
 Translations 🗣:
  -